rename rocq extraction pytest files
