+++ rocq-python-extraction/test/conftest.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 
 EXPLICIT_TARGETS = {
-    "check_core_terms_syntax.py": [
+    "test_core_terms_syntax.py": [
         "nat_add.py",
         "mk_pair_r.py",
         "zeros.py",
@@ -12,21 +12,21 @@ EXPLICIT_TARGETS = {
         "str_val.py",
         "todo_val.py",
     ],
-    "check_coinductives.py": [
+    "test_coinductives.py": [
         "repeat_tree.py",
         "tree_root_of_repeat.py",
         "zeros.py",
         "zeros_pair.py",
     ],
-    "check_modules.py": ["Phase10Mod.py"],
-    "check_point5.py": [
+    "test_modules.py": ["Phase10Mod.py"],
+    "test_point5.py": [
         "get_p5_v.py",
         "get_p5_w.py",
         "get_p5_x.py",
         "get_p5_y.py",
         "get_p5_z.py",
     ],
-    "check_proj_pair_r.py": [
+    "test_proj_pair_r.py": [
         "proj_first.py",
         "proj_second.py",
         "swap_pair_r.py",
@@ -38,12 +38,15 @@ def required_generated_files(path: Path) -> list[str]:
     explicit = EXPLICIT_TARGETS.get(path.name)
     if explicit is not None:
         return explicit
-    return [f"{path.stem.removeprefix('check_')}.py"]
+    return [f"{path.stem.removeprefix('test_')}.py"]
 
 
 def pytest_ignore_collect(collection_path: Path, config) -> bool:
     path = Path(str(collection_path))
-    if path.suffix != ".py" or not path.name.startswith("check_"):
+    if path.suffix != ".py" or not path.name.startswith("test_"):
+        return False
+
+    if path.name in {"test_support.py"}:
         return False
 
     repo_root = Path(__file__).resolve().parents[2]
