+++ rocq-python-extraction/test/test_coinductives.py
@@ -1,7 +1,7 @@
 # ruff: noqa: E402
 from itertools import islice
 
-from check_support import add_build_default_to_syspath, run_as_script
+from test_support import add_build_default_to_syspath, run_as_script
 
 add_build_default_to_syspath()
